+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Improvements:
 * The `type` of an `Email` is now non-null.
 * The `phoneNumber` of an `SMS` is now non-null.
 * The `latitude` and `longitude` of a `GeoPoint` are now non-null.
+* The `phones` and `urls` of `ContactInfo` are now non-null.
 
 ## 3.5.0
 New Features: